Update README.md
A Medical Clinic urgently needs a small application that allows us to manage the register of patient’s
appointments. To each patient appointment, who is registered, several medical exams may be associated.
The appointments and the exams are services provided by the Clinic, which are characterized with an
identification code (ID), the date of the appointment and its cost. Additionally, each appointment should
include a short text with its diagnosis, and each exam in which typology fits: Clinical Analysis, Imaging or
Check-up. In the patient's record should only be included the patient's name and his identification
number. Both the patient number and the service identification code must be sequential.
In addition to ensuring patient registration and appointments, the application should also allow
medical exams to be added to existing appointments. The application should also be able to determine
the total amount billed by the office and to show all the appointments of a given patient, presented all
relevant data of each appointment, including the number of associated exams.